small modif in viral data
an implementation error was present, but uneffective with this very specifiy dataset (basically, at the time to store the viral_df, we used a simple insert. This indeed conserves the order, but I guess items are simply dropped at the end of the Series if the two lengths do not match. In our case, the very last individual was the only one dropped, thus no error was present in downstream analyses, but this remains an implementation error